cleaning some files and adding a function to make **map variable.
