+++ main.c
@@ -6,16 +6,11 @@
 /*   By: mmarcott <mmarcott@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/26 15:00:32 by mmarcott          #+#    #+#             */
-/*   Updated: 2023/02/18 14:11:40 by mmarcott         ###   ########.fr       */
+/*   Updated: 2023/02/20 10:47:22 by mmarcott         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "so_long.h"
-
-void	ft_error(char *message)
-{
-	ft_printf("Error: %s\n", message);
-}
 
 void	place_decor(t_tile *decs, t_game *valeur)
 {
@@ -34,35 +29,6 @@ void	place_decor(t_tile *decs, t_game *valeur)
 		if (current->next == NULL)
 			return ;
 		current = current->next;
-	}
-}
-
-int	check_list(t_tile *head, int x, int y)
-{
-	t_tile	*current;
-
-	current = head;
-	while (current->next)
-	{
-		if ((current->type == 's') && x == current->x && y == current->y)
-			return (1);
-		current = current->next;
-	}
-	return (0);
-}
-
-void	move_player(t_game *game, int xmodifier, int ymodifier)
-{
-	if (!check_list(game->decors, game->player->x + xmodifier, game->player->y
-			+ ymodifier))
-	{
-		mlx_put_image_to_window(game->mlx, game->win, game->player->img,
-				game->player->x += xmodifier, game->player->y += ymodifier);
-	}
-	else
-	{
-		mlx_put_image_to_window(game->mlx, game->win, game->player->img,
-				game->player->x, game->player->y);
 	}
 }
 
@@ -121,24 +87,6 @@ void	add_end_list(t_tile **head, t_tile *ajout)
 	current->next = ajout;
 }
 
-t_tile	*init_decor(t_game *game, char *img, char type)
-{
-	t_tile	*new;
-
-	new = ft_calloc(1, sizeof(t_tile));
-	new->img = mlx_xpm_file_to_image(game->mlx, img, &new->x, &new->y);
-	new->y = 0;
-	new->x = 0;
-	if (type == 'c' || type == 'w')
-	{
-		new->x = 100;
-		new->y = 100;
-	}
-	new->type = type;
-	new->next = NULL;
-	return (new);
-}
-
 int	main(int argc, char **argv)
 {
 	char	*map;
@@ -154,11 +102,10 @@ int	main(int argc, char **argv)
 	ft_check_tiles_value(open(map, O_RDONLY));
 	game->title = "Le chien marin";
 	game->decors = ft_calloc(1, sizeof(t_tile));
-	ft_assign_tiles(&game, open(map, O_RDONLY));
 	game->mlx = mlx_init();
 	game->win = mlx_new_window(game->mlx, game->width, game->height,
 			game->title);
-	mlx_hook(game->win, 2, 0, key_pressed, game);
+	//mlx_hook(game->win, 2, 0, key_pressed, game);
 	mlx_loop(game->mlx);
 	return (0);
 }
